+++ src/config/creditLines/details/tabs.config.tsx
