+++ src/config/creditLines/table.config.tsx
@@ -1,14 +1,14 @@
-import { MdDeleteOutline } from "react-icons/md";
+import { MdDeleteOutline, MdOutlineCreate } from "react-icons/md";
 import { Icon } from "@inubekit/icon";
 
-import { IAction, ITitle } from "@components/data/Table/types";
+import { IAction, IEntry, ITitle } from "@components/data/Table/types";
 import { ComponentAppearance } from "@ptypes/aparences.types";
-import { Details } from "../components/Details";
+import { Details } from "@pages/creditLines/components/Details";
 
 const titles: ITitle[] = [
   {
-    id: "date",
-    titleName: "Fecha",
+    id: "name",
+    titleName: "Nombre",
     priority: 0,
   },
   {
@@ -16,19 +16,25 @@ const titles: ITitle[] = [
     titleName: "Descripción",
     priority: 1,
   },
-  {
-    id: "status",
-    titleName: "Estado",
-    priority: 2,
-  },
 ];
 
 const actions: IAction[] = [
   {
     id: "Details",
-    content: (entry) => <Details data={entry} />,
-  },
 
+    content: (entry: IEntry) => <Details data={entry} />,
+  },
+  {
+    id: "edit",
+    content: () => (
+      <Icon
+        appearance={ComponentAppearance.PRIMARY}
+        icon={<MdOutlineCreate />}
+        size="16px"
+        cursorHover
+      />
+    ),
+  },
   {
     id: "delete",
     content: () => (
@@ -43,7 +49,7 @@ const actions: IAction[] = [
 ];
 
 const breakPoints = [
-  { breakpoint: "(min-width: 745px)", totalColumns: 3 },
+  { breakpoint: "(min-width: 745px)", totalColumns: 2 },
   { breakpoint: "(max-width: 744px)", totalColumns: 1 },
 ];
