+++ src/config/moneyDestination/addDestination/assisted.config.ts
