+++ src/config/moneyDestination/moneyDestinationTab/generics/delete/messages.config.tsx
