+++ src/config/moneyDestination/requestsInProgressTab/table.config.tsx
@@ -1,14 +1,14 @@
-import { MdDeleteOutline, MdOutlineCreate } from "react-icons/md";
+import { MdDeleteOutline } from "react-icons/md";
 import { Icon } from "@inubekit/icon";
 
-import { IAction, IEntry, ITitle } from "@components/data/Table/types";
+import { IAction, ITitle } from "@components/data/Table/types";
 import { ComponentAppearance } from "@ptypes/aparences.types";
-import { Details } from "../components/Details";
+import { Details } from "@pages/moneyDestination/tabs/requestsInProgressTab/components/details";
 
 const titles: ITitle[] = [
   {
-    id: "name",
-    titleName: "Nombre",
+    id: "date",
+    titleName: "Fecha",
     priority: 0,
   },
   {
@@ -16,25 +16,19 @@ const titles: ITitle[] = [
     titleName: "Descripción",
     priority: 1,
   },
+  {
+    id: "status",
+    titleName: "Estado",
+    priority: 2,
+  },
 ];
 
 const actions: IAction[] = [
   {
     id: "Details",
+    content: (entry) => <Details data={entry} />,
+  },
 
-    content: (entry: IEntry) => <Details data={entry} />,
-  },
-  {
-    id: "edit",
-    content: () => (
-      <Icon
-        appearance={ComponentAppearance.PRIMARY}
-        icon={<MdOutlineCreate />}
-        size="16px"
-        cursorHover
-      />
-    ),
-  },
   {
     id: "delete",
     content: () => (
@@ -49,7 +43,7 @@ const actions: IAction[] = [
 ];
 
 const breakPoints = [
-  { breakpoint: "(min-width: 745px)", totalColumns: 2 },
+  { breakpoint: "(min-width: 745px)", totalColumns: 3 },
   { breakpoint: "(max-width: 744px)", totalColumns: 1 },
 ];
