+++ src/config/moneyDestination/tabs.config.tsx
