+++ src/pages/creditLines/addCreditLine/index.tsx
@@ -1,6 +1,7 @@
 import { useAddCreditlines } from "@hooks/creditLine/useAddCreditLines";
-import { addCreditLinesSteps } from "./config/assisted.config";
+
 import { AddCreditLinesUI } from "./interface";
+import { addCreditLinesSteps } from "@config/creditLines/addCreditLine/assisted.config";
 
 function Addcreditlines() {
   const {
