+++ src/pages/moneyDestination/tabs/moneyDestinationTab/components/delete/styles.ts
