+++ src/pages/moneyDestination/tabs/moneyDestinationTab/components/details/styles.ts
