+++ src/pages/moneyDestination/tabs/moneyDestinationTab/components/details/tabs.config.tsx
