+++ src/pages/moneyDestination/tabs/moneyDestinationTab/components/edit/index.tsx
@@ -1,20 +1,21 @@
-import { MdDeleteOutline } from "react-icons/md";
+import { MdOutlineCreate } from "react-icons/md";
 import { useState } from "react";
 import { Icon } from "@inubekit/icon";
+import { useMediaQuery } from "@inubekit/hooks";
+import { Text } from "@inubekit/text";
 
 import { DecisionModal } from "@components/modals/DecisionModal";
 import { ComponentAppearance } from "@ptypes/aparences.types";
 import { IEntry } from "@components/data/Table/types";
-import { deleteMoneyDestinationModal } from "../Delete/config/messages.config";
-import { useMediaQuery } from "@inubekit/hooks";
-import { StyledContainerIcon } from "./styles";
-import { Text } from "@inubekit/text";
 
-interface IDelete {
+import { StyledContainerIcon } from "./styles";
+import { editMoneyDestinationModal } from "@config/moneyDestination/moneyDestinationTab/generics/edit/messages.config";
+
+interface IEdit {
   data: IEntry;
 }
 
-const Delete = (props: IDelete) => {
+const Edit = (props: IEdit) => {
   const { data } = props;
   const [showModal, setShowModal] = useState(false);
 
@@ -28,26 +29,25 @@ const Delete = (props: IDelete) => {
     <>
       <StyledContainerIcon onClick={handleToggleModal} $isTablet={screenTablet}>
         <Icon
-          appearance={ComponentAppearance.DANGER}
-          icon={<MdDeleteOutline />}
-          size="16px"
-          onClick={handleToggleModal}
+          appearance={ComponentAppearance.PRIMARY}
+          icon={<MdOutlineCreate />}
+          size={screenTablet ? "20px" : "16px"}
           cursorHover
           spacing="narrow"
         />
         {screenTablet && (
           <Text type="body" size="medium">
-            Eliminar
+            Editar
           </Text>
         )}
       </StyledContainerIcon>
+
       {showModal && (
         <DecisionModal
           portalId="portal"
-          title={deleteMoneyDestinationModal.title}
-          actionText={deleteMoneyDestinationModal.actionText}
-          description={deleteMoneyDestinationModal.description}
-          justificationOfDecision={true}
+          title={editMoneyDestinationModal.title}
+          actionText={editMoneyDestinationModal.actionText}
+          description={editMoneyDestinationModal.description}
           onClick={() => console.log(data)}
           onCloseModal={handleToggleModal}
         />
@@ -56,4 +56,4 @@ const Delete = (props: IDelete) => {
   );
 };
 
-export { Delete };
+export { Edit };
