+++ src/pages/moneyDestination/tabs/moneyDestinationTab/forms/verificationForm/index.tsx
@@ -11,8 +11,8 @@ import { IRequestSteps } from "@design/feedback/RequestProcess/types";
 import { RequestProcessModal } from "@components/modals/RequestProcessModal";
 import { VerificationBoxes } from "./verificationBoxes";
 import { IFormsUpdateData } from "./types";
-import { addDestinationStepsConfig } from "../../addDestination/config/assisted.config";
-import { finishModal } from "./config/messages.config";
+import { addDestinationStepsConfig } from "@config/moneyDestination/addDestination/assisted.config";
+import { finishModal } from "@config/moneyDestination/moneyDestinationTab/form/verificationForm/messages.config";
 
 interface IVerificationForm {
   requestSteps: IRequestSteps[];
