+++ src/pages/moneyDestination/tabs/requestsInProgressTab/components/details/styles.ts
