+++ src/pages/moneyDestination/tabs/requestsInProgressTab/interface.tsx
@@ -6,7 +6,7 @@ import { tokens } from "@design/tokens";
 import { Table } from "@components/data/Table";
 import { dataRequestsInProgress } from "@mocks/moneydestination/requestsInProgress.mock";
 import { StyledContainer } from "./styles";
-import { actions, breakPoints, titles } from "./config/table.config";
+import { actions, breakPoints, titles } from "@config/moneyDestination/requestsInProgressTab/table.config";
 
 interface IRequestsInProgressTabUI {
   loading: boolean;
